Move new interfaces/functions to the root folder.
